+++ src/PhpPresentation/Writer/PowerPoint2007/AbstractDecoratorWriter.php
@@ -2,6 +2,7 @@
 
 namespace PhpOffice\PhpPresentation\Writer\PowerPoint2007;
 
+use PhpOffice\Common\Adapter\Zip\ZipInterface;
 use PhpOffice\Common\Drawing as CommonDrawing;
 use PhpOffice\Common\XMLWriter;
 use PhpOffice\PhpPresentation\HashTable;
@@ -14,7 +15,7 @@ use \ZipArchive;
 abstract class AbstractDecoratorWriter
 {
     /**
-     * @return \ZipArchive
+     * @return ZipInterface
      */
     abstract public function render();
 
@@ -72,18 +73,18 @@ abstract class AbstractDecoratorWriter
     }
 
     /**
-     * @var ZipArchive
+     * @var ZipInterface
      */
     protected $oZip;
 
-    public function setZip(ZipArchive $oZip)
+    public function setZip(ZipInterface $oZip)
     {
         $this->oZip = $oZip;
         return $this;
     }
 
     /**
-     * @return ZipArchive
+     * @return ZipInterface
      */
     public function getZip()
     {
@@ -335,6 +336,7 @@ abstract class AbstractDecoratorWriter
         // > a:ln
         $objWriter->endElement();
     }
+
     /**
      * Determine absolute zip path
      *