+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - ODPresentation & PowerPoint2007 Writer : Thumbnail of the presentation - @Progi1984 GH-125
 - ODPresentation & PowerPoint2007 Writer : Marker of Series in Line & Scatter chart is customizable - @Progi1984 GH-169
 - ODPresentation & PowerPoint2007 Writer : Outline of Series in Line & Scatter chart is customizable - @Progi1984 GH-169
+- PowerPoint2007 / Serialized Writer : Support for Zip Adapter - @Progi1984 GH-176
 - ODPresentation & PowerPoint2007 Writer : Add Font Support For Chart Axis - @jrking4 GH-186
 
 ## 0.6.0 - 2016-01-24